+++ ndonnx/_core/_numericimpl.py
@@ -22,7 +22,6 @@ from ndonnx._utility import promote
 from ._coreimpl import CoreOperationsImpl
 from ._interface import OperationsBlock
 from ._nullableimpl import NullableOperationsImpl
-from ._shapeimpl import UniformShapeOperations
 from ._utils import (
     binary_op,
     from_corearray,
@@ -971,14 +970,10 @@ class _NumericOperationsImpl(OperationsBlock):
         return ndx.full_like(x, 0, dtype=dtype)
 
 
-class NumericOperationsImpl(
-    CoreOperationsImpl, _NumericOperationsImpl, UniformShapeOperations
-): ...
+class NumericOperationsImpl(CoreOperationsImpl, _NumericOperationsImpl): ...
 
 
-class NullableNumericOperationsImpl(
-    NullableOperationsImpl, _NumericOperationsImpl, UniformShapeOperations
-): ...
+class NullableNumericOperationsImpl(NullableOperationsImpl, _NumericOperationsImpl): ...
 
 
 def _via_i64_f64(