+++ CHANGELOG.rst
@@ -22,6 +22,7 @@ Changelog
 - Various operations that depend on the array's shape have been updated to work correctly with lazy arrays.
 - :func:`ndonnx.cumulative_sum` now correctly applies the ``include_initial`` parameter and works around missing onnxruntime kernels for unsigned integral types.
 - :func:`ndonnx.additional.make_nullable` applies broadcasting to the provided null array (instead of reshape like it did previously). This allows writing ``make_nullable(x, False)`` to turn an array into nullable.
+- User-defined data types that implement :class:`ndonnx._core.UniformShapeOperations` may now implement :func:`ndonnx.where` without requiring both data types be promotable.
 
 **Breaking change**